a lot of things

+++ src/main/java/ru/ravel/ItDesk/dto/OrganizationPriorityDuration.java
@@ -8,5 +8,5 @@ import ru.ravel.ItDesk.model.Priority;
 public class OrganizationPriorityDuration {
 	private Organization organization;
 	private Priority priority;
-	private Long hours;
+	private Float hours;
 }
